Improved vision processing functions
Made the autonomousUpdate function require a filtered image to be
passed to it and created a separate filterImage function that is used
for filtering the image before passing it to the autonomousUpdate.
There also is a filterImageTest that saves the images as there are
filtered in the cRIO flash memory.